+++ src/main/java/com/mnb/service/PublisherService.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface PublisherService {
     public List<Publisher> findAll();
 
-  public Publisher findById(int theId);
+  public Publisher findById(long theId);
 
     public void save(Publisher thePublisher);
 
-    public void deleteById(int theId);
+    public void deleteById(long theId);
 }
